[MLOP-225] KeyFeature need dtype to be a required arg (#126)

+++ tests/integration/butterfree/core/pipelines/test_feature_set_pipeline.py
@@ -104,7 +104,11 @@ class TestFeatureSetPipeline:
                     ),
                 ],
                 keys=[
-                    KeyFeature(name="id", description="The user's Main ID or device ID")
+                    KeyFeature(
+                        name="id",
+                        description="The user's Main ID or device ID",
+                        dtype=DataType.INTEGER,
+                    )
                 ],
                 timestamp=TimestampFeature(),
             ),
